+++ core/dbt/artifacts/resources/v1/components.py
@@ -1,6 +1,10 @@
 from dataclasses import dataclass, field
 from dbt.artifacts.resources.v1.macro import MacroDependsOn
-from typing import List
+from dbt_common.dataclass_schema import dbtClassMixin
+from typing import Dict, List, Optional, Union
+
+
+NodeVersion = Union[str, float]
 
 
 @dataclass
@@ -10,3 +14,24 @@ class DependsOn(MacroDependsOn):
     def add_node(self, value: str):
         if value not in self.nodes:
             self.nodes.append(value)
+
+
+@dataclass
+class RefArgs(dbtClassMixin):
+    name: str
+    package: Optional[str] = None
+    version: Optional[NodeVersion] = None
+
+    @property
+    def positional_args(self) -> List[str]:
+        if self.package:
+            return [self.package, self.name]
+        else:
+            return [self.name]
+
+    @property
+    def keyword_args(self) -> Dict[str, Optional[NodeVersion]]:
+        if self.version:
+            return {"version": self.version}
+        else:
+            return {}
