+++ core/dbt/contracts/graph/node_args.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 from datetime import datetime
 from typing import Optional, List
 
-from dbt.contracts.graph.unparsed import NodeVersion
+from dbt.artifacts.resources import NodeVersion
 from dbt.node_types import NodeType, AccessType
