+++ core/dbt/plugins/manifest.py
@@ -6,7 +6,7 @@ from dbt.contracts.graph.node_args import ModelNodeArgs
 # all these are just exports, they need "noqa" so flake8 will not complain.
 from dbt.contracts.graph.manifest import Manifest  # noqa
 from dbt.node_types import AccessType, NodeType  # noqa
-from dbt.contracts.graph.unparsed import NodeVersion  # noqa
+from dbt.artifacts.resources import NodeVersion  # noqa
 from dbt.graph.graph import UniqueId  # noqa
