+++ tests/functional/experimental_parser/test_all_experimental_parser.py
@@ -2,8 +2,8 @@ import pytest
 
 from dbt.tests.util import run_dbt, run_dbt_and_capture
 
+from dbt.artifacts.resources import RefArgs
 from dbt.contracts.graph.manifest import Manifest
-from dbt.contracts.graph.nodes import RefArgs
 import os
