+++ tests/unit/test_contracts_graph_parsed.py
@@ -4,6 +4,7 @@ import pytest
 from hypothesis import given
 from hypothesis.strategies import builds, lists
 
+from dbt.artifacts.resources import RefArgs
 from dbt.node_types import NodeType, AccessType
 from dbt.contracts.files import FileHash
 from dbt.contracts.graph.model_config import (
@@ -38,7 +39,6 @@ from dbt.contracts.graph.nodes import (
     Owner,
     TestMetadata,
     SemanticModel,
-    RefArgs,
 )
 from dbt.contracts.graph.semantic_models import Dimension, Entity, Measure
 from dbt.contracts.graph.unparsed import (
